Added a MainMenuActivity
I've added a button to MainActivity that opens MainMenuActivity.

In main menu activity, I have implemented a very basic settings bar
that's blank for the moment. The main menu has 3 buttons that don't do
anything at the moment but the code is implemented so that they can open
new activities. I just have it commented out since the classes for the
dictionary, word history and the map have not been created yet.